fix: detect namespaces readyed by third party contracts

+++ src/migrations/1608030374841_namespaces.ts
@@ -33,11 +33,11 @@ export async function up(pgm: MigrationBuilder): Promise<void> {
       notNull: true,
     },
     base: {
-      type: 'integer',
+      type: 'numeric',
       notNull: true,
     },
     coeff: {
-      type: 'integer',
+      type: 'numeric',
       notNull: true,
     },
     nonalpha_discount: {
